tx payment refactor - with_normalized moved back to TxPayment
