+++ framework/base/src/types/interaction/tx_payment/tx_payment_multi_esdt.rs
@@ -1,6 +1,6 @@
 use crate::{
     contract_base::SendRawWrapper,
-    types::{BigUint, ManagedAddress, ManagedVec, MultiEsdtPayment},
+    types::{BigUint, ManagedAddress, ManagedVec, MultiEsdtPayment, TxFrom, TxToSpecified},
 };
 
 use super::{AnnotatedEgldPayment, FullPaymentData, FunctionCall, TxEgldValue, TxEnv, TxPayment};
@@ -29,6 +29,29 @@ where
         );
     }
 
+    fn with_normalized<From, To, F, R>(
+        self,
+        env: &Env,
+        from: &From,
+        to: To,
+        fc: FunctionCall<Env::Api>,
+        f: F,
+    ) -> R
+    where
+        From: TxFrom<Env>,
+        To: TxToSpecified<Env>,
+        F: FnOnce(&ManagedAddress<Env::Api>, &BigUint<Env::Api>, &FunctionCall<Env::Api>) -> R,
+    {
+        match self.len() {
+            0 => ().with_normalized(env, from, to, fc, f),
+            1 => self.get(0).with_normalized(env, from, to, fc, f),
+            _ => to.with_address_ref(env, |to_addr| {
+                let fc_conv = fc.convert_to_multi_transfer_esdt_call(to_addr, self);
+                f(&from.resolve_address(env), &BigUint::zero(), &fc_conv)
+            }),
+        }
+    }
+
     fn into_full_payment_data(self, _env: &Env) -> FullPaymentData<Env::Api> {
         FullPaymentData {
             egld: None,
@@ -53,6 +76,22 @@ where
         fc: FunctionCall<Env::Api>,
     ) {
         (&self).perform_transfer_execute(env, to, gas_limit, fc);
+    }
+
+    fn with_normalized<From, To, F, R>(
+        self,
+        env: &Env,
+        from: &From,
+        to: To,
+        fc: FunctionCall<Env::Api>,
+        f: F,
+    ) -> R
+    where
+        From: TxFrom<Env>,
+        To: TxToSpecified<Env>,
+        F: FnOnce(&ManagedAddress<Env::Api>, &BigUint<Env::Api>, &FunctionCall<Env::Api>) -> R,
+    {
+        (&self).with_normalized(env, from, to, fc, f)
     }
 
     fn into_full_payment_data(self, _env: &Env) -> FullPaymentData<Env::Api> {
