+++ framework/base/src/types/interaction/tx_payment/tx_payment_none.rs
@@ -1,6 +1,6 @@
 use crate::{
     contract_base::SendRawWrapper,
-    types::{BigUint, ManagedAddress, ManagedVec},
+    types::{BigUint, ManagedAddress, ManagedVec, TxFrom, TxToSpecified},
 };
 
 use super::{
@@ -24,6 +24,22 @@ where
         fc: FunctionCall<Env::Api>,
     ) {
         Egld(BigUint::zero()).perform_transfer_execute(env, to, gas_limit, fc);
+    }
+
+    fn with_normalized<From, To, F, R>(
+        self,
+        env: &Env,
+        _from: &From,
+        to: To,
+        fc: FunctionCall<Env::Api>,
+        f: F,
+    ) -> R
+    where
+        From: TxFrom<Env>,
+        To: TxToSpecified<Env>,
+        F: FnOnce(&ManagedAddress<Env::Api>, &BigUint<Env::Api>, &FunctionCall<Env::Api>) -> R,
+    {
+        to.with_address_ref(env, |to_addr| f(to_addr, &BigUint::zero(), &fc))
     }
 
     fn into_full_payment_data(self, _env: &Env) -> FullPaymentData<Env::Api> {
