+++ framework/base/src/types/interaction/tx_payment/tx_payment_other.rs
@@ -2,6 +2,7 @@ use crate::{
     contract_base::SendRawWrapper,
     types::{
         BigUint, Egld, EgldOrEsdtTokenPayment, EgldOrMultiEsdtPayment, ManagedAddress, ManagedVec,
+        TxFrom, TxToSpecified,
     },
 };
 
@@ -26,6 +27,26 @@ where
             (to, fc),
             |(to, fc), amount| Egld(amount).perform_transfer_execute(env, to, gas_limit, fc),
             |(to, fc), esdt_payment| esdt_payment.perform_transfer_execute(env, to, gas_limit, fc),
+        )
+    }
+
+    fn with_normalized<From, To, F, R>(
+        self,
+        env: &Env,
+        from: &From,
+        to: To,
+        fc: FunctionCall<Env::Api>,
+        f: F,
+    ) -> R
+    where
+        From: TxFrom<Env>,
+        To: TxToSpecified<Env>,
+        F: FnOnce(&ManagedAddress<Env::Api>, &BigUint<Env::Api>, &FunctionCall<Env::Api>) -> R,
+    {
+        self.map_egld_or_esdt(
+            (to, fc, f),
+            |(to, fc, f), amount| Egld(amount).with_normalized(env, from, to, fc, f),
+            |(to, fc, f), esdt_payment| esdt_payment.with_normalized(env, from, to, fc, f),
         )
     }
 
@@ -59,6 +80,29 @@ where
             },
             EgldOrMultiEsdtPayment::MultiEsdt(multi_esdt_payment) => {
                 multi_esdt_payment.perform_transfer_execute(env, to, gas_limit, fc)
+            },
+        }
+    }
+
+    fn with_normalized<From, To, F, R>(
+        self,
+        env: &Env,
+        from: &From,
+        to: To,
+        fc: FunctionCall<Env::Api>,
+        f: F,
+    ) -> R
+    where
+        From: TxFrom<Env>,
+        To: TxToSpecified<Env>,
+        F: FnOnce(&ManagedAddress<Env::Api>, &BigUint<Env::Api>, &FunctionCall<Env::Api>) -> R,
+    {
+        match self {
+            EgldOrMultiEsdtPayment::Egld(egld_amount) => {
+                Egld(egld_amount).with_normalized(env, from, to, fc, f)
+            },
+            EgldOrMultiEsdtPayment::MultiEsdt(multi_esdt_payment) => {
+                multi_esdt_payment.with_normalized(env, from, to, fc, f)
             },
         }
     }
